fix(ci): resolve failing PHP linting and Psalm checks
- fix: align array double arrow in test-class-sensei-question.php
- fix: cast $lesson_id to int in class-sensei-question.php to match
  output_result_indication()'s declared int param

Errors fixed:
- WordPress.Arrays.MultipleStatementAlignment.DoubleArrowNotAligned
- Psalm InvalidScalarArgument

PHP 8.2 compatible. Milestone and artifact-upload CI checks left
unresolved: milestone needs a maintainer with write access, artifact
upload is a token-permissions issue on the base repo.

Refs #8074

+++ includes/class-sensei-question.php
@@ -1113,7 +1113,7 @@ class Sensei_Question {
 
 		$quiz_id            = $sensei_question_loop['quiz_id'];
 		$question_item      = $sensei_question_loop['current_question'];
-		$lesson_id          = Sensei()->quiz->get_lesson_id( $quiz_id );
+		$lesson_id          = (int) Sensei()->quiz->get_lesson_id( $quiz_id );
 		$user_lesson_status = Sensei_Utils::user_lesson_status( $lesson_id, get_current_user_id() );
 		$quiz_graded        = isset( $user_lesson_status->comment_approved ) && ! in_array( $user_lesson_status->comment_approved, array( 'in-progress', 'ungraded' ) );
